+++ changelog.md
@@ -4,3 +4,4 @@
 - Remove `database.init_db`.
 - Modify column `idf` to real number.
 - Handle Twitter error on answer.
+- Add `after_death` command for tweet after you died.